requirements.txt
git+https://github.com/plamere/spotipy.git
vk_api
colorama
python-dotenv
